Changed Add Problems
Add problems now includes a button that links to the different types of forums that the admin will choose from.
